攻撃中断で回避するときに、入力方向を反映するように改善

+++ Src/Object/Character/Enemy/EnemyOnHit.cpp
@@ -59,15 +59,20 @@ void EnemyOnHit::CollNormalAttack(const std::weak_ptr<Collider> _hitCol)
 	if (weapon.GetIsDamage()||charaObj_.GetCardAction()==CardActionBase::CARD_ACT_TYPE::JUMP_ATK)return;
 	//ダメージを与えたことを知らせる
 	weapon.SetIsDamage();
+
 	//相手キャラの技攻撃力取得
 	const float& atkPoint = action_.GetInput().GetTargetCharacter().lock()->GetMainAction().GetAtkStatus().atkPoint;
+
 	//相手の数字との差が近いほどダメージを追加する
 	const int cardDif= CardSystem::GetInstance().GetCardDif();
-	float addDam = 0;
+
+	int addDam = 0;
 	//カードシステムからカードの強さの差を取得し、０でなければダメージを追加する
-	cardDif==0?addDam=0:addDam= MAX_ADD_DAMAGE - cardDif;
-	float damage = atkPoint + addDam;
+	cardDif == 0 ? addDam = 0 : addDam = MAX_ADD_DAMAGE - cardDif;
+
+	int damage = atkPoint + addDam;
 	charaObj_.Damage(damage);
+
 	//charaObj_.ChangeUpdatePhase(CharacterBase::UPDATE_PHASE::HIT_STOP);
 	action_.ChangeAction(ActionController::ACTION_TYPE::REACT);
 }
